refactor(tests): split test_utils.py into a focused utils package (#443)
test_utils.py was 1453 LOC covering four unrelated subsystems
(teatree.utils.ports, .git, .db and teatree.backends.gitlab_api),
making it slow to open and hard to navigate.

Pure verbatim reorganization mirroring the production layout into
tests/utils/: each subsystem gets its own focused module, with the
large gitlab_api surface further split by client/queries/auth-cache
concern. Test bodies are AST-identical to the original; only import
paths changed. No production code touched, no assertions changed.

Also refreshes the stale mirror example in tests/README.md: the
sync.py example now points at a module that still maps 1:1
(resolve.py) and notes the package convention for oversized modules.

Relates-to #443
