--- tests/test_utils.py
@@ -1,1453 +1,0 @@
-import os
-from pathlib import Path
-from subprocess import CompletedProcess
-from unittest.mock import patch
-
-import pytest
-
-from teatree.backends import gitlab_api
-from teatree.utils import db, git, ports
-from teatree.utils import run as utils_run_mod
-
-
-def test_get_service_port_parses_docker_compose_output(monkeypatch: pytest.MonkeyPatch) -> None:
-    """get_service_port parses `docker compose port` output."""
-    monkeypatch.setattr(
-        utils_run_mod.subprocess,
-        "run",
-        lambda *_a, **_k: CompletedProcess([], 0, stdout="0.0.0.0:8042\n"),
-    )
-    assert ports.get_service_port("myproject", "web", 8000) == 8042
-
-
-def test_get_service_port_returns_none_on_failure(monkeypatch: pytest.MonkeyPatch) -> None:
-    """get_service_port returns None when service is not running."""
-    monkeypatch.setattr(
-        utils_run_mod.subprocess,
-        "run",
-        lambda *_a, **_k: CompletedProcess([], 1, stdout=""),
-    )
-    assert ports.get_service_port("myproject", "web", 8000) is None
-
-
-def test_get_worktree_ports_queries_all_services(monkeypatch: pytest.MonkeyPatch) -> None:
-    """get_worktree_ports queries all compose services and returns named ports."""
-    port_map = {
-        ("web", 8000): "0.0.0.0:8042\n",
-        ("frontend", 80): "0.0.0.0:4242\n",
-        ("db", 5432): "",  # not running
-    }
-
-    def fake_run(cmd: list[str], **kwargs: object) -> CompletedProcess[str]:
-        service = cmd[-2]
-        container_port = int(cmd[-1])
-        key = (service, container_port)
-        output = port_map.get(key, "")
-        return CompletedProcess(cmd, 0 if output else 1, stdout=output)
-
-    monkeypatch.setattr(utils_run_mod.subprocess, "run", fake_run)
-
-    result = ports.get_worktree_ports("myproject")
-    assert result == {"backend": 8042, "frontend": 4242}
-    assert "postgres" not in result  # db service was not running
-
-
-def test_default_branch_prefers_symbolic_ref_and_falls_back(monkeypatch: pytest.MonkeyPatch) -> None:
-    calls: list[list[str]] = []
-
-    def fake_run(
-        args: list[str],
-        *,
-        capture_output: bool,
-        text: bool = False,
-        check: bool = False,
-        **_kwargs: object,
-    ) -> CompletedProcess[str]:
-        calls.append(args)
-        if args[-1] == "refs/remotes/origin/HEAD":
-            return CompletedProcess(args, 1, "", "")
-        if args[-1] == "refs/remotes/origin/main":
-            return CompletedProcess(args, 0, "", "")
-        return CompletedProcess(args, 1, "", "")
-
-    monkeypatch.setattr(utils_run_mod.subprocess, "run", fake_run)
-
-    assert git.default_branch("/tmp/repo") == "main"
-    assert calls[0][-1] == "refs/remotes/origin/HEAD"
-
-
-def test_default_branch_returns_symbolic_ref_when_present(monkeypatch: pytest.MonkeyPatch) -> None:
-    monkeypatch.setattr(
-        utils_run_mod.subprocess,
-        "run",
-        lambda *args, **kwargs: CompletedProcess(list(args[0]), 0, "refs/remotes/origin/main\n", ""),
-    )
-
-    assert git.default_branch("/tmp/repo") == "main"
-
-
-def test_git_helpers_cover_run_check_current_branch_and_failure(monkeypatch: pytest.MonkeyPatch) -> None:
-    def fake_run(
-        args: list[str],
-        *,
-        capture_output: bool,
-        text: bool = False,
-        check: bool = False,
-        **_kwargs: object,
-    ) -> CompletedProcess[str]:
-        if args[-2:] == ["status", "--short"]:
-            return CompletedProcess(args, 0, " M pyproject.toml\n", "")
-        if args[-2:] == ["rev-parse", "--abbrev-ref"]:
-            return CompletedProcess(args, 0, "", "")
-        return CompletedProcess(args, 1, "", "")
-
-    monkeypatch.setattr(utils_run_mod.subprocess, "run", fake_run)
-
-    assert git.run(repo="/tmp/repo", args=["status", "--short"]) == "M pyproject.toml"
-    assert git.check(repo="/tmp/repo", args=["status", "--short"]) is True
-    assert git.current_branch("/tmp/repo") == ""
-    with pytest.raises(RuntimeError, match="Could not detect default branch"):
-        git.default_branch("/tmp/repo")
-
-
-def test_run_strict_raises_on_failure(monkeypatch: pytest.MonkeyPatch) -> None:
-    def fake_run(
-        args: list[str],
-        *,
-        capture_output: bool,
-        text: bool = False,
-        check: bool = False,
-        **_: object,
-    ) -> CompletedProcess[str]:
-        if "bad" in args:
-            return CompletedProcess(args, 1, "", "fatal")
-        return CompletedProcess(args, 0, "ok\n", "")
-
-    monkeypatch.setattr(utils_run_mod.subprocess, "run", fake_run)
-
-    assert git.run(repo="/tmp/r", args=["log"]) == "ok"
-    with pytest.raises(utils_run_mod.CommandFailedError):
-        git.run_strict(repo="/tmp/r", args=["bad"])
-
-
-def test_git_high_level_operations(monkeypatch: pytest.MonkeyPatch) -> None:
-    calls: list[list[str]] = []
-
-    def fake_run(
-        args: list[str],
-        *,
-        capture_output: bool,
-        text: bool = False,
-        check: bool = False,
-        **_kwargs: object,
-    ) -> CompletedProcess[str]:
-        calls.append(list(args))
-        if "merge-base" in args:
-            return CompletedProcess(args, 0, "abc123\n", "")
-        if "rev-list" in args:
-            return CompletedProcess(args, 0, "3\n", "")
-        if "log" in args:
-            return CompletedProcess(args, 0, "abc feat one\ndef feat two\n", "")
-        if "status" in args:
-            return CompletedProcess(args, 0, " M file.py\n", "")
-        return CompletedProcess(args, 0, "", "")
-
-    monkeypatch.setattr(utils_run_mod.subprocess, "run", fake_run)
-
-    assert git.merge_base("/tmp/r", "origin/main") == "abc123"
-    assert git.rev_count("/tmp/r", "abc123..HEAD") == 3
-    assert git.log_oneline("/tmp/r", "abc123..HEAD") == "abc feat one\ndef feat two"
-    assert git.status_porcelain("/tmp/r") == "M file.py"
-
-    git.soft_reset("/tmp/r", "abc123")
-    assert any("reset" in c for c in calls)
-
-    git.commit("/tmp/r", "squash msg")
-    assert any("commit" in c for c in calls)
-
-    git.fetch("/tmp/r", "origin", "main")
-    assert any("fetch" in c for c in calls)
-
-    git.rebase("/tmp/r", "origin/main")
-    assert any("rebase" in c for c in calls)
-
-
-def test_git_worktree_and_branch_ops(monkeypatch: pytest.MonkeyPatch) -> None:
-    def fake_run(
-        args: list[str],
-        *,
-        capture_output: bool,
-        text: bool = False,
-        check: bool = False,
-        **_kwargs: object,
-    ) -> CompletedProcess[str]:
-        if "worktree" in args:
-            return CompletedProcess(args, 0, "", "")
-        if "branch" in args:
-            return CompletedProcess(args, 1, "", "")
-        if "pull" in args:
-            return CompletedProcess(args, 0, "", "")
-        return CompletedProcess(args, 0, "", "")
-
-    monkeypatch.setattr(utils_run_mod.subprocess, "run", fake_run)
-
-    assert git.worktree_remove("/tmp/r", "/tmp/wt") is True
-    assert git.branch_delete("/tmp/r", "old-branch") is False
-    assert git.pull_ff_only("/tmp/r") is True
-
-
-def test_fetch_without_ref(monkeypatch: pytest.MonkeyPatch) -> None:
-    calls: list[list[str]] = []
-
-    def fake_run(
-        args: list[str],
-        *,
-        capture_output: bool,
-        text: bool = False,
-        check: bool = False,
-        **_kwargs: object,
-    ) -> CompletedProcess[str]:
-        calls.append(list(args))
-        return CompletedProcess(args, 0, "", "")
-
-    monkeypatch.setattr(utils_run_mod.subprocess, "run", fake_run)
-
-    git.fetch("/tmp/r")
-    assert calls[-1] == ["git", "-C", "/tmp/r", "fetch", "origin"]
-
-
-def test_remote_url_returns_url(monkeypatch: pytest.MonkeyPatch) -> None:
-    monkeypatch.setattr(
-        utils_run_mod.subprocess,
-        "run",
-        lambda *a, **kw: CompletedProcess(a[0], 0, "git@github.com:acme/repo.git\n", ""),
-    )
-    assert git.remote_url(repo="/tmp/r", remote="origin") == "git@github.com:acme/repo.git"
-
-
-@pytest.mark.parametrize(
-    ("remote_url_value", "expected_slug"),
-    [
-        ("git@github.com:acme/repo.git", "acme/repo"),
-        ("git@github.com:acme/repo", "acme/repo"),
-        ("https://github.com/acme/repo.git", "acme/repo"),
-        ("https://github.com/acme/repo", "acme/repo"),
-        ("ssh://git@github.com/acme/repo.git", "acme/repo"),
-        ("git@gitlab.com:group/sub/proj.git", "group/sub/proj"),
-        ("https://gitlab.com/group/sub/proj.git", "group/sub/proj"),
-    ],
-)
-def test_remote_slug_parses_supported_url_forms(
-    remote_url_value: str,
-    expected_slug: str,
-    monkeypatch: pytest.MonkeyPatch,
-) -> None:
-    monkeypatch.setattr(
-        utils_run_mod.subprocess,
-        "run",
-        lambda *a, **kw: CompletedProcess(a[0], 0, f"{remote_url_value}\n", ""),
-    )
-    assert git.remote_slug(repo="/tmp/r", remote="origin") == expected_slug
-
-
-def test_remote_slug_returns_empty_when_no_remote(monkeypatch: pytest.MonkeyPatch) -> None:
-    monkeypatch.setattr(
-        utils_run_mod.subprocess,
-        "run",
-        lambda *a, **kw: CompletedProcess(a[0], 1, "", "no remote"),
-    )
-    assert git.remote_slug(repo="/tmp/r") == ""
-
-
-def test_remote_slug_passes_through_when_path_is_already_slug(monkeypatch: pytest.MonkeyPatch) -> None:
-    """Callers can hand an already-resolved slug (``owner/repo``) and get it back unchanged."""
-    assert git.remote_slug(repo="acme/repo") == "acme/repo"
-    assert git.remote_slug(repo="group/sub/proj") == "group/sub/proj"
-
-
-def test_config_value_returns_value(monkeypatch: pytest.MonkeyPatch) -> None:
-    monkeypatch.setattr(
-        utils_run_mod.subprocess,
-        "run",
-        lambda *a, **kw: CompletedProcess(a[0], 0, "Jane Doe\n", ""),
-    )
-    assert git.config_value(key="user.name") == "Jane Doe"
-
-
-def test_last_commit_message_parses_subject_and_body(monkeypatch: pytest.MonkeyPatch) -> None:
-    monkeypatch.setattr(
-        utils_run_mod.subprocess,
-        "run",
-        lambda *a, **kw: CompletedProcess(a[0], 0, "fix: bug\n\nDetailed body\n", ""),
-    )
-    subject, body = git.last_commit_message(repo="/tmp/r")
-    assert subject == "fix: bug"
-    assert body == "Detailed body"
-
-
-def test_last_commit_message_subject_only(monkeypatch: pytest.MonkeyPatch) -> None:
-    monkeypatch.setattr(
-        utils_run_mod.subprocess,
-        "run",
-        lambda *a, **kw: CompletedProcess(a[0], 0, "feat: add feature", ""),
-    )
-    subject, body = git.last_commit_message()
-    assert subject == "feat: add feature"
-    assert body == ""
-
-
-def test_worktree_add_with_and_without_create_branch(monkeypatch: pytest.MonkeyPatch) -> None:
-    calls: list[list[str]] = []
-
-    def fake_run(
-        args: list[str],
-        *,
-        capture_output: bool,
-        text: bool = False,
-        check: bool = False,
-        **_kwargs: object,
-    ) -> CompletedProcess[str]:
-        calls.append(list(args))
-        return CompletedProcess(args, 0, "", "")
-
-    monkeypatch.setattr(utils_run_mod.subprocess, "run", fake_run)
-
-    assert git.worktree_add("/tmp/r", "/tmp/wt", "feat-1", create_branch=True) is True
-    assert "-b" in calls[-1]
-
-    assert git.worktree_add("/tmp/r", "/tmp/wt2", "feat-1", create_branch=False) is True
-    assert "-b" not in calls[-1]
-    assert "feat-1" in calls[-1]
-
-
-def test_db_restore_uses_pg_restore_when_supported(monkeypatch: pytest.MonkeyPatch) -> None:
-    commands: list[list[str]] = []
-    monkeypatch.setenv("POSTGRES_HOST", "db.internal")
-    monkeypatch.setenv("POSTGRES_USER", "postgres")
-    monkeypatch.setenv("POSTGRES_PASSWORD", "secret")
-
-    def fake_run(
-        args: list[str],
-        *,
-        env: dict[str, str] | None = None,
-        capture_output: bool = False,
-        text: bool = False,
-        check: bool = False,
-        **_kwargs: object,
-    ) -> CompletedProcess[str]:
-        commands.append(args)
-        if args[:2] == ["pg_restore", "-l"]:
-            return CompletedProcess(args, 0, "toc", "")
-        if args[0] == "pg_restore":
-            return CompletedProcess(args, 0, "", "")
-        return CompletedProcess(args, 0, "", "")
-
-    monkeypatch.setattr(utils_run_mod.subprocess, "run", fake_run)
-
-    db.db_restore("wt_123", "/tmp/dump.dump")
-
-    assert commands[0][:2] == ["dropdb", "-h"]
-    assert commands[1][:2] == ["createdb", "-h"]
-    assert commands[2] == ["pg_restore", "-l", "/tmp/dump.dump"]
-    assert commands[3][0] == "pg_restore"
-
-
-def test_db_helpers_cover_env_exists_and_psql_fallback(monkeypatch: pytest.MonkeyPatch) -> None:
-    monkeypatch.delenv("POSTGRES_PASSWORD", raising=False)
-    monkeypatch.delenv("POSTGRES_PORT", raising=False)
-    monkeypatch.setenv("POSTGRES_HOST", "db.internal")
-    monkeypatch.setenv("POSTGRES_USER", "worker")
-
-    commands: list[list[str]] = []
-
-    def fake_run(
-        args: list[str],
-        *,
-        env: dict[str, str] | None = None,
-        capture_output: bool = False,
-        text: bool = False,
-        check: bool = False,
-        **_kwargs: object,
-    ) -> CompletedProcess[str]:
-        commands.append(args)
-        if args[:2] == ["pg_restore", "-l"]:
-            return CompletedProcess(args, 1, "", "")
-        if args[0] == "psql" and "-lqt" in args:
-            return CompletedProcess(args, 0, "wt_42 | owner\n", "")
-        return CompletedProcess(args, 0, "", "")
-
-    monkeypatch.setattr(utils_run_mod.subprocess, "run", fake_run)
-
-    assert db.pg_env().get("PGPASSWORD") is None
-    assert db.pg_host() == "db.internal"
-    assert db.pg_user() == "worker"
-    assert db.worktree_db_name("42", "") == "wt_42"
-    db.db_restore("wt_42", "/tmp/dump.sql")
-    assert db.db_exists("wt_42") is True
-    assert commands[3][0] == "psql"
-
-
-def test_db_restore_raises_when_restore_commands_fail(monkeypatch: pytest.MonkeyPatch) -> None:
-    def fake_run(
-        args: list[str],
-        *,
-        env: dict[str, str] | None = None,
-        capture_output: bool = False,
-        text: bool = False,
-        check: bool = False,
-        **_kwargs: object,
-    ) -> CompletedProcess[str]:
-        if args[:2] == ["pg_restore", "-l"]:
-            return CompletedProcess(args, 0, "toc", "")
-        if args[0] == "pg_restore":
-            return CompletedProcess(args, 1, "", "")
-        if args[0] == "psql":
-            return CompletedProcess(args, 1, "", "")
-        return CompletedProcess(args, 0, "", "")
-
-    monkeypatch.setattr(utils_run_mod.subprocess, "run", fake_run)
-
-    with pytest.raises(RuntimeError, match="pg_restore failed"):
-        db.db_restore("wt_55", "/tmp/dump.dump")
-
-
-def test_db_restore_raises_when_psql_restore_fails(monkeypatch: pytest.MonkeyPatch) -> None:
-    def fake_run(
-        args: list[str],
-        *,
-        env: dict[str, str] | None = None,
-        capture_output: bool = False,
-        text: bool = False,
-        check: bool = False,
-        **_kwargs: object,
-    ) -> CompletedProcess[str]:
-        if args[:2] == ["pg_restore", "-l"]:
-            return CompletedProcess(args, 1, "", "")
-        if args[0] == "psql":
-            return CompletedProcess(args, 1, "", "")
-        return CompletedProcess(args, 0, "", "")
-
-    monkeypatch.setattr(utils_run_mod.subprocess, "run", fake_run)
-
-    with pytest.raises(RuntimeError, match="psql restore failed"):
-        db.db_restore("wt_56", "/tmp/dump.sql")
-
-
-def test_db_restore_detects_truncated_pg_restore(monkeypatch: pytest.MonkeyPatch) -> None:
-    def fake_run(
-        args: list[str],
-        *,
-        env: dict[str, str] | None = None,
-        capture_output: bool = False,
-        text: bool = False,
-        check: bool = False,
-        **_kwargs: object,
-    ) -> CompletedProcess[str]:
-        if args[:2] == ["pg_restore", "-l"]:
-            return CompletedProcess(args, 0, "TOC", "")
-        if args[0] == "pg_restore" and "-d" in args:
-            return CompletedProcess(args, 0, "", "WARNING: could not read data")
-        return CompletedProcess(args, 0, "", "")
-
-    monkeypatch.setattr(utils_run_mod.subprocess, "run", fake_run)
-
-    with pytest.raises(RuntimeError, match="Corrupt or truncated dump"):
-        db.db_restore("wt_70", "/tmp/dump.pgdump")
-
-
-def test_db_restore_detects_truncated_psql(monkeypatch: pytest.MonkeyPatch) -> None:
-    def fake_run(
-        args: list[str],
-        *,
-        env: dict[str, str] | None = None,
-        capture_output: bool = False,
-        text: bool = False,
-        check: bool = False,
-        **_kwargs: object,
-    ) -> CompletedProcess[str]:
-        if args[:2] == ["pg_restore", "-l"]:
-            return CompletedProcess(args, 1, "", "")
-        if args[0] == "psql":
-            return CompletedProcess(args, 0, "", "unexpected EOF on client connection")
-        return CompletedProcess(args, 0, "", "")
-
-    monkeypatch.setattr(utils_run_mod.subprocess, "run", fake_run)
-
-    with pytest.raises(RuntimeError, match="Corrupt or truncated dump"):
-        db.db_restore("wt_71", "/tmp/dump.sql")
-
-
-def test_gitlab_api_resolves_remote_project(monkeypatch: pytest.MonkeyPatch) -> None:
-    monkeypatch.setattr(git, "remote_url", lambda **kwargs: "git@gitlab.com:acme/platform.git")
-
-    client = gitlab_api.GitLabAPI(token="test-token")
-    monkeypatch.setattr(
-        client,
-        "get_json",
-        lambda endpoint: {
-            "id": 42,
-            "path_with_namespace": "acme/platform",
-            "path": "platform",
-        },
-    )
-
-    project = client.resolve_project_from_remote("/tmp/repo")
-
-    assert project is not None
-    assert project.project_id == 42
-    assert project.short_name == "platform"
-
-
-def test_gitlab_api_helpers_cover_http_paths_and_failures(monkeypatch: pytest.MonkeyPatch) -> None:
-    requests: list[str] = []
-
-    def fake_get(
-        url: str,
-        *,
-        headers: dict[str, str],
-        timeout: float,
-    ) -> object:
-        requests.append(url)
-
-        class Response:
-            def raise_for_status(self) -> None:
-                return None
-
-            def json(self) -> dict[str, object] | list[dict[str, object]]:
-                if url.endswith("projects/acme%2Fplatform"):
-                    return {"id": 42, "path_with_namespace": "acme/platform", "path": "platform"}
-                if "merge_requests" in url:
-                    return [{"iid": 1, "draft": False}, {"iid": 2, "draft": True}]
-                return [{"id": 101}]
-
-        return Response()
-
-    def fake_post(
-        url: str,
-        *,
-        headers: dict[str, str],
-        json: dict[str, object],
-        timeout: float,
-    ) -> object:
-        requests.append(url)
-
-        class Response:
-            def raise_for_status(self) -> None:
-                return None
-
-            def json(self) -> dict[str, object]:
-                return {"ok": True}
-
-        return Response()
-
-    monkeypatch.setattr(gitlab_api.httpx, "get", fake_get)
-    monkeypatch.setattr(gitlab_api.httpx, "post", fake_post)
-    monkeypatch.setattr(git, "remote_url", lambda **kwargs: "")
-    monkeypatch.setattr(git, "current_branch", lambda **kwargs: "")
-
-    client = gitlab_api.GitLabAPI(token="")
-    assert client.get_json("projects/x") is None
-    assert client.post_json("projects/x") is None
-
-    client = gitlab_api.GitLabAPI(token="test-token")
-    assert client.resolve_project("acme/platform") == client.resolve_project("acme/platform")
-    assert client.resolve_project("missing/project") is None
-    assert client.list_all_open_mrs("adrien") == [{"iid": 1, "draft": False}, {"iid": 2, "draft": True}]
-    assert client.list_all_open_mrs("adrien", include_draft=False) == [{"iid": 1, "draft": False}]
-    assert client.cancel_pipelines(42, "feature") == [101, 101]
-    monkeypatch.setattr(client, "get_json", lambda endpoint: None)
-    assert client.list_all_open_mrs("adrien") == []
-    monkeypatch.setattr(client, "get_json", lambda endpoint: {"oops": "bad"})
-    assert client.cancel_pipelines(42, "feature") == []
-    assert client.resolve_project_from_remote("/tmp/repo") is None
-    assert gitlab_api.GitLabAPI.current_branch("/tmp/repo") == ""
-    assert requests
-
-
-def test_gitlab_api_returns_none_for_non_gitlab_remote(monkeypatch: pytest.MonkeyPatch) -> None:
-    monkeypatch.setattr(git, "remote_url", lambda **kwargs: "https://example.com/acme/repo.git")
-
-    assert gitlab_api.GitLabAPI(token="test-token").resolve_project_from_remote("/tmp/repo") is None
-
-
-def test_pg_env_includes_port_when_set(monkeypatch: pytest.MonkeyPatch) -> None:
-    monkeypatch.setenv("POSTGRES_PORT", "5433")
-    monkeypatch.setenv("POSTGRES_PASSWORD", "secret")
-
-    env = db.pg_env()
-
-    assert env["PGPORT"] == "5433"
-    assert env["PGPASSWORD"] == "secret"
-
-
-def test_pg_env_omits_port_when_unset(monkeypatch: pytest.MonkeyPatch) -> None:
-    monkeypatch.delenv("POSTGRES_PORT", raising=False)
-    monkeypatch.delenv("POSTGRES_PASSWORD", raising=False)
-
-    env = db.pg_env()
-
-    assert "PGPORT" not in env
-
-
-def test_pg_env_resolves_password_via_pass_key(monkeypatch: pytest.MonkeyPatch) -> None:
-    """``pg_env`` reads the password from ``pass`` when ``POSTGRES_PASSWORD_PASS_KEY`` is set.
-
-    The literal value never appears in ``os.environ`` — the only secret
-    delivery channel is the in-memory ``PGPASSWORD`` value placed on the
-    returned ``env`` dict.
-    """
-    from unittest.mock import patch  # noqa: PLC0415
-
-    monkeypatch.delenv("POSTGRES_PASSWORD", raising=False)
-    monkeypatch.setenv("POSTGRES_PASSWORD_PASS_KEY", "teatree/wt/42/postgres")
-    with patch("teatree.utils.postgres_secret.secrets.read_pass", return_value="from-pass"):
-        env = db.pg_env()
-    assert env["PGPASSWORD"] == "from-pass"
-    # The literal value must not be planted into the original process env.
-    assert os.environ.get("POSTGRES_PASSWORD", "") == ""
-
-
-def test_gitlab_api_graphql_sends_post_request(monkeypatch: pytest.MonkeyPatch) -> None:
-    posted: list[dict[str, object]] = []
-
-    def fake_post(
-        url: str,
-        *,
-        headers: dict[str, str],
-        json: dict[str, object],
-        timeout: float,
-    ) -> object:
-        posted.append({"url": url, "json": json})
-
-        class Response:
-            def raise_for_status(self) -> None:
-                return None
-
-            def json(self) -> dict[str, object]:
-                return {"data": {"project": {"workItems": {"nodes": []}}}}
-
-        return Response()
-
-    monkeypatch.setattr(gitlab_api.httpx, "post", fake_post)
-
-    client = gitlab_api.GitLabAPI(token="test-token")
-    result = client.graphql("query { project { id } }", {"projectPath": "org/repo"})
-
-    assert result is not None
-    assert result["data"] is not None
-    assert posted[0]["url"] == "https://gitlab.com/api/graphql"
-
-
-def test_gitlab_api_graphql_returns_none_without_token() -> None:
-    client = gitlab_api.GitLabAPI(token="")
-
-    result = client.graphql("query { project { id } }")
-
-    assert result is None
-
-
-def test_get_work_item_status_returns_status_name(monkeypatch: pytest.MonkeyPatch) -> None:
-    client = gitlab_api.GitLabAPI(token="test-token")
-    monkeypatch.setattr(
-        client,
-        "graphql",
-        lambda query, variables: {
-            "data": {
-                "project": {
-                    "workItems": {
-                        "nodes": [
-                            {
-                                "widgets": [
-                                    {"type": "DESCRIPTION"},
-                                    {"type": "STATUS", "status": {"name": "In Progress"}},
-                                ],
-                            },
-                        ],
-                    },
-                },
-            },
-        },
-    )
-
-    result = client.get_work_item_status("org/repo", 42)
-
-    assert result == "In Progress"
-
-
-def test_get_work_item_status_returns_none_when_graphql_fails(monkeypatch: pytest.MonkeyPatch) -> None:
-    client = gitlab_api.GitLabAPI(token="test-token")
-    monkeypatch.setattr(client, "graphql", lambda query, variables: None)
-
-    result = client.get_work_item_status("org/repo", 42)
-
-    assert result is None
-
-
-def test_get_work_item_status_returns_none_for_empty_nodes(monkeypatch: pytest.MonkeyPatch) -> None:
-    client = gitlab_api.GitLabAPI(token="test-token")
-    monkeypatch.setattr(
-        client,
-        "graphql",
-        lambda query, variables: {
-            "data": {
-                "project": {
-                    "workItems": {
-                        "nodes": [],
-                    },
-                },
-            },
-        },
-    )
-
-    result = client.get_work_item_status("org/repo", 42)
-
-    assert result is None
-
-
-def test_get_work_item_status_returns_none_when_no_status_widget(monkeypatch: pytest.MonkeyPatch) -> None:
-    client = gitlab_api.GitLabAPI(token="test-token")
-    monkeypatch.setattr(
-        client,
-        "graphql",
-        lambda query, variables: {
-            "data": {
-                "project": {
-                    "workItems": {
-                        "nodes": [
-                            {
-                                "widgets": [
-                                    {"type": "DESCRIPTION"},
-                                ],
-                            },
-                        ],
-                    },
-                },
-            },
-        },
-    )
-
-    result = client.get_work_item_status("org/repo", 42)
-
-    assert result is None
-
-
-def test_get_work_item_status_returns_none_when_widgets_not_a_list(monkeypatch: pytest.MonkeyPatch) -> None:
-    client = gitlab_api.GitLabAPI(token="test-token")
-    monkeypatch.setattr(
-        client,
-        "graphql",
-        lambda query, variables: {
-            "data": {
-                "project": {
-                    "workItems": {
-                        "nodes": [
-                            {
-                                "widgets": "not-a-list",
-                            },
-                        ],
-                    },
-                },
-            },
-        },
-    )
-
-    result = client.get_work_item_status("org/repo", 42)
-
-    assert result is None
-
-
-def test_get_work_item_status_returns_none_when_status_value_not_dict(monkeypatch: pytest.MonkeyPatch) -> None:
-    client = gitlab_api.GitLabAPI(token="test-token")
-    monkeypatch.setattr(
-        client,
-        "graphql",
-        lambda query, variables: {
-            "data": {
-                "project": {
-                    "workItems": {
-                        "nodes": [
-                            {
-                                "widgets": [
-                                    {"type": "STATUS", "status": "not-a-dict"},
-                                ],
-                            },
-                        ],
-                    },
-                },
-            },
-        },
-    )
-
-    result = client.get_work_item_status("org/repo", 42)
-
-    assert result is None
-
-
-def test_list_all_open_mrs_with_updated_after(monkeypatch: pytest.MonkeyPatch) -> None:
-    client = gitlab_api.GitLabAPI(token="test-token")
-    monkeypatch.setattr(
-        client,
-        "get_json",
-        lambda endpoint: [{"iid": 1, "draft": False}],
-    )
-
-    result = client.list_all_open_mrs("adrien", updated_after="2024-01-01T00:00:00Z")
-
-    assert result == [{"iid": 1, "draft": False}]
-
-
-def test_list_open_mrs_as_reviewer(monkeypatch: pytest.MonkeyPatch) -> None:
-    client = gitlab_api.GitLabAPI(token="test-token")
-    captured_endpoints: list[str] = []
-
-    def _capture(endpoint: str) -> list[dict[str, object]]:
-        captured_endpoints.append(endpoint)
-        return [{"iid": 5, "web_url": "https://gitlab.com/org/repo/-/merge_requests/5"}]
-
-    monkeypatch.setattr(client, "get_json", _capture)
-
-    result = client.list_open_mrs_as_reviewer("adrien")
-
-    assert result == [{"iid": 5, "web_url": "https://gitlab.com/org/repo/-/merge_requests/5"}]
-    assert "reviewer_username=adrien" in captured_endpoints[0]
-    assert "not%5Bauthor_username%5D=adrien" in captured_endpoints[0]
-
-
-def test_list_open_issues_for_assignee(monkeypatch: pytest.MonkeyPatch) -> None:
-    client = gitlab_api.GitLabAPI(token="test-token")
-    captured_endpoints: list[str] = []
-
-    def _capture(endpoint: str) -> list[dict[str, object]]:
-        captured_endpoints.append(endpoint)
-        return [{"iid": 42, "web_url": "https://gitlab.com/org/repo/-/issues/42"}]
-
-    monkeypatch.setattr(client, "get_json", _capture)
-
-    result = client.list_open_issues_for_assignee("adrien", updated_after="2024-01-01T00:00:00Z")
-
-    assert result == [{"iid": 42, "web_url": "https://gitlab.com/org/repo/-/issues/42"}]
-    assert captured_endpoints[0].startswith("issues?")
-    assert "assignee_username=adrien" in captured_endpoints[0]
-    assert "state=opened" in captured_endpoints[0]
-    assert "updated_after=2024-01-01T00%3A00%3A00Z" in captured_endpoints[0]
-
-
-def test_list_open_issues_for_assignee_returns_empty_on_non_list(monkeypatch: pytest.MonkeyPatch) -> None:
-    client = gitlab_api.GitLabAPI(token="test-token")
-    monkeypatch.setattr(client, "get_json", lambda _endpoint: None)
-
-    assert client.list_open_issues_for_assignee("adrien") == []
-
-
-def test_list_recently_merged_mrs_returns_data(monkeypatch: pytest.MonkeyPatch) -> None:
-    client = gitlab_api.GitLabAPI(token="test-token")
-    monkeypatch.setattr(
-        client,
-        "get_json",
-        lambda endpoint: [{"iid": 10, "state": "merged"}],
-    )
-
-    result = client.list_recently_merged_mrs("adrien")
-
-    assert result == [{"iid": 10, "state": "merged"}]
-
-
-def test_list_recently_merged_mrs_with_updated_after(monkeypatch: pytest.MonkeyPatch) -> None:
-    client = gitlab_api.GitLabAPI(token="test-token")
-    captured_endpoints: list[str] = []
-
-    def capture_get_json(endpoint: str) -> list[dict[str, object]]:
-        captured_endpoints.append(endpoint)
-        return [{"iid": 10}]
-
-    monkeypatch.setattr(client, "get_json", capture_get_json)
-
-    result = client.list_recently_merged_mrs("adrien", updated_after="2024-06-01T00:00:00Z")
-
-    assert result == [{"iid": 10}]
-    assert "updated_after" in captured_endpoints[0]
-
-
-def test_list_recently_merged_mrs_returns_empty_when_not_a_list(monkeypatch: pytest.MonkeyPatch) -> None:
-    client = gitlab_api.GitLabAPI(token="test-token")
-    monkeypatch.setattr(client, "get_json", lambda endpoint: {"error": "bad"})
-
-    result = client.list_recently_merged_mrs("adrien")
-
-    assert result == []
-
-
-def test_list_recently_closed_mrs_queries_state_closed(monkeypatch: pytest.MonkeyPatch) -> None:
-    client = gitlab_api.GitLabAPI(token="test-token")
-    captured: list[str] = []
-
-    def capture_get_json(endpoint: str) -> list[dict[str, object]]:
-        captured.append(endpoint)
-        return [{"iid": 77, "state": "closed"}]
-
-    monkeypatch.setattr(client, "get_json", capture_get_json)
-
-    result = client.list_recently_closed_mrs("adrien", updated_after="2024-06-01T00:00:00Z")
-
-    assert result == [{"iid": 77, "state": "closed"}]
-    assert "state=closed" in captured[0]
-    assert "author_username=adrien" in captured[0]
-    assert "updated_after" in captured[0]
-
-
-def test_list_recently_closed_mrs_returns_empty_when_not_a_list(monkeypatch: pytest.MonkeyPatch) -> None:
-    client = gitlab_api.GitLabAPI(token="test-token")
-    monkeypatch.setattr(client, "get_json", lambda _endpoint: {"error": "bad"})
-
-    assert client.list_recently_closed_mrs("adrien") == []
-
-
-def test_get_mr_pipeline_returns_status_and_url(monkeypatch: pytest.MonkeyPatch) -> None:
-    client = gitlab_api.GitLabAPI(token="test-token")
-    monkeypatch.setattr(
-        client,
-        "get_json",
-        lambda endpoint: [{"status": "success", "web_url": "https://gitlab.com/pipelines/1"}],
-    )
-
-    result = client.get_mr_pipeline(42, 1)
-
-    assert result == {"status": "success", "url": "https://gitlab.com/pipelines/1"}
-
-
-def test_get_mr_pipeline_returns_none_when_no_pipelines(monkeypatch: pytest.MonkeyPatch) -> None:
-    client = gitlab_api.GitLabAPI(token="test-token")
-    monkeypatch.setattr(client, "get_json", lambda endpoint: [])
-
-    result = client.get_mr_pipeline(42, 1)
-
-    assert result == {"status": None, "url": None}
-
-
-def test_get_mr_pipeline_returns_none_when_not_a_list(monkeypatch: pytest.MonkeyPatch) -> None:
-    client = gitlab_api.GitLabAPI(token="test-token")
-    monkeypatch.setattr(client, "get_json", lambda endpoint: None)
-
-    result = client.get_mr_pipeline(42, 1)
-
-    assert result == {"status": None, "url": None}
-
-
-def test_get_mr_approvals_returns_counts(monkeypatch: pytest.MonkeyPatch) -> None:
-    client = gitlab_api.GitLabAPI(token="test-token")
-    monkeypatch.setattr(
-        client,
-        "get_json",
-        lambda endpoint: {
-            "approved_by": [{"user": {"username": "reviewer1"}}, {"user": {"username": "reviewer2"}}],
-            "approvals_required": 2,
-        },
-    )
-
-    result = client.get_mr_approvals(42, 1)
-
-    assert result == {"count": 2, "required": 2, "approved_by": ["reviewer1", "reviewer2"]}
-
-
-def test_get_mr_approvals_returns_defaults_when_not_a_dict(monkeypatch: pytest.MonkeyPatch) -> None:
-    client = gitlab_api.GitLabAPI(token="test-token")
-    monkeypatch.setattr(client, "get_json", lambda endpoint: None)
-
-    result = client.get_mr_approvals(42, 1)
-
-    assert result == {"count": 0, "required": 1, "approved_by": []}
-
-
-def test_get_mr_approvals_handles_non_list_approved_by(monkeypatch: pytest.MonkeyPatch) -> None:
-    client = gitlab_api.GitLabAPI(token="test-token")
-    monkeypatch.setattr(
-        client,
-        "get_json",
-        lambda endpoint: {"approved_by": "not-a-list", "approvals_required": 1},
-    )
-
-    result = client.get_mr_approvals(42, 1)
-
-    assert result == {"count": 0, "required": 1, "approved_by": []}
-
-
-def test_get_mr_approvals_skips_non_dict_entry(monkeypatch: pytest.MonkeyPatch) -> None:
-    """get_mr_approvals skips non-dict entries in approved_by list (line 213)."""
-    client = gitlab_api.GitLabAPI(token="test-token")
-    monkeypatch.setattr(
-        client,
-        "get_json",
-        lambda endpoint: {
-            "approved_by": [
-                "not-a-dict",
-                {"user": {"username": "reviewer1"}},
-            ],
-            "approvals_required": 1,
-        },
-    )
-
-    result = client.get_mr_approvals(42, 1)
-
-    assert result == {"count": 2, "required": 1, "approved_by": ["reviewer1"]}
-
-
-def test_get_issue_returns_issue_dict(monkeypatch: pytest.MonkeyPatch) -> None:
-    client = gitlab_api.GitLabAPI(token="test-token")
-    monkeypatch.setattr(
-        client,
-        "get_json",
-        lambda endpoint: {"iid": 5, "title": "Bug"},
-    )
-
-    result = client.get_issue(42, 5)
-
-    assert result == {"iid": 5, "title": "Bug"}
-
-
-def test_get_issue_returns_none_when_not_a_dict(monkeypatch: pytest.MonkeyPatch) -> None:
-    client = gitlab_api.GitLabAPI(token="test-token")
-    monkeypatch.setattr(client, "get_json", lambda endpoint: None)
-
-    result = client.get_issue(42, 5)
-
-    assert result is None
-
-
-def test_get_mr_discussions_returns_list(monkeypatch: pytest.MonkeyPatch) -> None:
-    client = gitlab_api.GitLabAPI(token="test-token")
-    monkeypatch.setattr(
-        client,
-        "get_json",
-        lambda endpoint: [{"id": "d1", "notes": []}],
-    )
-
-    result = client.get_mr_discussions(42, 1)
-
-    assert result == [{"id": "d1", "notes": []}]
-
-
-def test_get_mr_discussions_returns_empty_when_not_a_list(monkeypatch: pytest.MonkeyPatch) -> None:
-    client = gitlab_api.GitLabAPI(token="test-token")
-    monkeypatch.setattr(client, "get_json", lambda endpoint: None)
-
-    result = client.get_mr_discussions(42, 1)
-
-    assert result == []
-
-
-def test_get_draft_notes_count_returns_count(monkeypatch: pytest.MonkeyPatch) -> None:
-    client = gitlab_api.GitLabAPI(token="test-token")
-    monkeypatch.setattr(
-        client,
-        "get_json",
-        lambda endpoint: [{"id": 1}, {"id": 2}, {"id": 3}],
-    )
-
-    result = client.get_draft_notes_count(42, 1)
-
-    assert result == 3
-
-
-def test_get_draft_notes_count_returns_zero_when_not_a_list(monkeypatch: pytest.MonkeyPatch) -> None:
-    client = gitlab_api.GitLabAPI(token="test-token")
-    monkeypatch.setattr(client, "get_json", lambda endpoint: None)
-
-    result = client.get_draft_notes_count(42, 1)
-
-    assert result == 0
-
-
-def test_current_username_returns_username(monkeypatch: pytest.MonkeyPatch) -> None:
-    client = gitlab_api.GitLabAPI(token="test-token")
-    monkeypatch.setattr(
-        client,
-        "get_json",
-        lambda endpoint: {"username": "adrien"},
-    )
-
-    result = client.current_username()
-
-    assert result == "adrien"
-
-
-def test_current_username_returns_empty_when_not_a_dict(monkeypatch: pytest.MonkeyPatch) -> None:
-    client = gitlab_api.GitLabAPI(token="test-token")
-    monkeypatch.setattr(client, "get_json", lambda endpoint: None)
-
-    result = client.current_username()
-
-    assert result == ""
-
-
-def test_gitlab_api_reads_token_from_env(monkeypatch: pytest.MonkeyPatch) -> None:
-    monkeypatch.setenv("GITLAB_TOKEN", "env-token")
-
-    client = gitlab_api.GitLabAPI()
-
-    assert client.token == "env-token"
-
-
-def test_gitlab_api_explicit_token_overrides_env(monkeypatch: pytest.MonkeyPatch) -> None:
-    monkeypatch.setenv("GITLAB_TOKEN", "env-token")
-
-    client = gitlab_api.GitLabAPI(token="explicit-token")
-
-    assert client.token == "explicit-token"
-
-
-def test_resolve_token_falls_back_to_pass(monkeypatch: pytest.MonkeyPatch) -> None:
-    monkeypatch.delenv("GITLAB_TOKEN", raising=False)
-    with patch("teatree.utils.secrets.read_pass", return_value="pass-token"):
-        assert gitlab_api._resolve_token() == "pass-token"
-
-
-def test_resolve_token_returns_empty_when_pass_fails(monkeypatch: pytest.MonkeyPatch) -> None:
-    monkeypatch.delenv("GITLAB_TOKEN", raising=False)
-    with patch("teatree.utils.secrets.read_pass", return_value=""):
-        assert gitlab_api._resolve_token() == ""
-
-
-def test_resolve_token_prefers_env_over_pass(monkeypatch: pytest.MonkeyPatch) -> None:
-    monkeypatch.setenv("GITLAB_TOKEN", "env-token")
-    assert gitlab_api._resolve_token() == "env-token"
-
-
-class TestGitLabAPICacheHits:
-    """Verify cache-hit branches for all cached methods."""
-
-    def test_get_work_item_status_returns_cached(self, monkeypatch: pytest.MonkeyPatch) -> None:
-        client = gitlab_api.GitLabAPI(token="t")
-        graphql_calls: list[int] = []
-        monkeypatch.setattr(
-            client,
-            "graphql",
-            lambda q, v: (
-                graphql_calls.append(1)
-                or {
-                    "data": {
-                        "project": {
-                            "workItems": {
-                                "nodes": [{"widgets": [{"type": "STATUS", "status": {"name": "In progress"}}]}],
-                            },
-                        },
-                    },
-                }
-            ),
-        )
-        assert client.get_work_item_status("org/repo", 1) == "In progress"
-        assert client.get_work_item_status("org/repo", 1) == "In progress"
-        assert len(graphql_calls) == 1
-
-    def test_get_mr_pipeline_returns_cached(self, monkeypatch: pytest.MonkeyPatch) -> None:
-        client = gitlab_api.GitLabAPI(token="t")
-        calls = []
-        monkeypatch.setattr(
-            client,
-            "get_json",
-            lambda ep: calls.append(1) or [{"status": "success", "web_url": "https://ci/1"}],
-        )
-        client.get_mr_pipeline(1, 1)
-        result = client.get_mr_pipeline(1, 1)
-        assert result["status"] == "success"
-        assert len(calls) == 1
-
-    def test_get_mr_approvals_returns_cached(self, monkeypatch: pytest.MonkeyPatch) -> None:
-        client = gitlab_api.GitLabAPI(token="t")
-        calls = []
-        monkeypatch.setattr(
-            client,
-            "get_json",
-            lambda ep: calls.append(1) or {"approved_by": [], "approvals_required": 1},
-        )
-        client.get_mr_approvals(1, 1)
-        result = client.get_mr_approvals(1, 1)
-        assert result["count"] == 0
-        assert len(calls) == 1
-
-    def test_get_issue_returns_cached(self, monkeypatch: pytest.MonkeyPatch) -> None:
-        client = gitlab_api.GitLabAPI(token="t")
-        calls = []
-        monkeypatch.setattr(client, "get_json", lambda ep: calls.append(1) or {"iid": 5, "title": "Bug"})
-        client.get_issue(1, 5)
-        result = client.get_issue(1, 5)
-        assert result is not None
-        assert result["title"] == "Bug"
-        assert len(calls) == 1
-
-    def test_get_mr_discussions_returns_cached(self, monkeypatch: pytest.MonkeyPatch) -> None:
-        client = gitlab_api.GitLabAPI(token="t")
-        calls = []
-        monkeypatch.setattr(client, "get_json", lambda ep: calls.append(1) or [{"id": "d1"}])
-        client.get_mr_discussions(1, 1)
-        result = client.get_mr_discussions(1, 1)
-        assert result == [{"id": "d1"}]
-        assert len(calls) == 1
-
-    def test_get_draft_notes_count_returns_cached(self, monkeypatch: pytest.MonkeyPatch) -> None:
-        client = gitlab_api.GitLabAPI(token="t")
-        calls = []
-        monkeypatch.setattr(client, "get_json", lambda ep: calls.append(1) or [{"id": 1}, {"id": 2}])
-        client.get_draft_notes_count(1, 1)
-        result = client.get_draft_notes_count(1, 1)
-        assert result == 2
-        assert len(calls) == 1
-
-    def test_current_username_returns_cached(self, monkeypatch: pytest.MonkeyPatch) -> None:
-        client = gitlab_api.GitLabAPI(token="t")
-        calls = []
-        monkeypatch.setattr(client, "get_json", lambda ep: calls.append(1) or {"username": "dev"})
-        client.current_username()
-        result = client.current_username()
-        assert result == "dev"
-        assert len(calls) == 1
-
-    def test_clear_response_cache(self, monkeypatch: pytest.MonkeyPatch) -> None:
-        client = gitlab_api.GitLabAPI(token="t")
-        calls = []
-        monkeypatch.setattr(client, "get_json", lambda ep: calls.append(1) or {"username": "dev"})
-        client.current_username()
-        client.clear_response_cache()
-        client.current_username()
-        assert len(calls) == 2
-
-
-class TestUnsyncedCommits:
-    def test_returns_empty_list_when_fully_synced(self, monkeypatch: pytest.MonkeyPatch) -> None:
-        monkeypatch.setattr(
-            utils_run_mod.subprocess,
-            "run",
-            lambda *_a, **_k: CompletedProcess([], 0, stdout="", stderr=""),
-        )
-        assert git.unsynced_commits("/repo", "feature") == []
-
-    def test_returns_commit_lines_when_commits_exist(self, monkeypatch: pytest.MonkeyPatch) -> None:
-        output = "abc123 chore: cve fix\ndef456 feat: add something\n"
-        monkeypatch.setattr(
-            utils_run_mod.subprocess,
-            "run",
-            lambda *_a, **_k: CompletedProcess([], 0, stdout=output, stderr=""),
-        )
-        result = git.unsynced_commits("/repo", "feature")
-        assert result == ["abc123 chore: cve fix", "def456 feat: add something"]
-
-    def test_filters_blank_lines_from_output(self, monkeypatch: pytest.MonkeyPatch) -> None:
-        output = "abc123 fix something\n\n   \ndef456 another fix\n"
-        monkeypatch.setattr(
-            utils_run_mod.subprocess,
-            "run",
-            lambda *_a, **_k: CompletedProcess([], 0, stdout=output, stderr=""),
-        )
-        result = git.unsynced_commits("/repo", "feature")
-        assert result == ["abc123 fix something", "def456 another fix"]
-
-    def test_pushed_branch_with_commit_not_on_main_is_ahead(self, tmp_path: Path) -> None:
-        """A commit pushed to its own remote branch but not on main must still be ahead.
-
-        Regression: ``git log --not --remotes`` excludes ALL remote tracking
-        refs, so a pushed feature branch's own commits get filtered out and
-        the classifier mis-reports the branch as synced. The unsynced check
-        must compare against the default branch, not against every remote.
-        """
-        bare = tmp_path / "remote.git"
-        utils_run_mod.run_checked(["git", "init", "--bare", str(bare)])
-        local = tmp_path / "local"
-        utils_run_mod.run_checked(["git", "clone", str(bare), str(local)])
-        for k, v in {"user.email": "t@x", "user.name": "t", "commit.gpgsign": "false"}.items():
-            utils_run_mod.run_checked(["git", "-C", str(local), "config", k, v])
-        (local / "a").write_text("1\n")
-        utils_run_mod.run_checked(["git", "-C", str(local), "add", "a"])
-        utils_run_mod.run_checked(["git", "-C", str(local), "commit", "-m", "main commit"])
-        utils_run_mod.run_checked(["git", "-C", str(local), "branch", "-M", "main"])
-        utils_run_mod.run_checked(["git", "-C", str(local), "push", "origin", "main"])
-        utils_run_mod.run_checked(["git", "-C", str(local), "checkout", "-b", "feature"])
-        (local / "b").write_text("2\n")
-        utils_run_mod.run_checked(["git", "-C", str(local), "add", "b"])
-        utils_run_mod.run_checked(["git", "-C", str(local), "commit", "-m", "feature work"])
-        utils_run_mod.run_checked(["git", "-C", str(local), "push", "origin", "feature"])
-
-        result = git.unsynced_commits(str(local), "feature")
-        assert len(result) == 1
-        assert "feature work" in result[0]
-
-
-class TestCommitsAbsentFromAllRemotes:
-    """#706 data-loss guard helper — commits reachable from no remote ref."""
-
-    def test_returns_empty_when_nothing_unpushed(self, monkeypatch: pytest.MonkeyPatch) -> None:
-        monkeypatch.setattr(
-            utils_run_mod.subprocess,
-            "run",
-            lambda *_a, **_k: CompletedProcess([], 0, stdout="", stderr=""),
-        )
-        assert git.commits_absent_from_all_remotes("/repo", "feature") == []
-
-    def test_returns_lines_and_filters_blanks(self, monkeypatch: pytest.MonkeyPatch) -> None:
-        output = "abc123 unpushed work\n\n   \ndef456 more local\n"
-        monkeypatch.setattr(
-            utils_run_mod.subprocess,
-            "run",
-            lambda *_a, **_k: CompletedProcess([], 0, stdout=output, stderr=""),
-        )
-        assert git.commits_absent_from_all_remotes("/repo", "feature") == [
-            "abc123 unpushed work",
-            "def456 more local",
-        ]
-
-    def test_pushed_branch_not_on_main_is_considered_safe(self, tmp_path: Path) -> None:
-        """A pushed-but-unmerged branch has nothing absent from all remotes.
-
-        The work survives on its own remote ref, so teardown is safe. This is
-        the inverse of ``unsynced_commits`` and the reason #706 needs a
-        distinct helper.
-        """
-        bare = tmp_path / "remote.git"
-        utils_run_mod.run_checked(["git", "init", "--bare", str(bare)])
-        local = tmp_path / "local"
-        utils_run_mod.run_checked(["git", "clone", str(bare), str(local)])
-        for k, v in {"user.email": "t@x", "user.name": "t", "commit.gpgsign": "false"}.items():
-            utils_run_mod.run_checked(["git", "-C", str(local), "config", k, v])
-        (local / "a").write_text("1\n")
-        utils_run_mod.run_checked(["git", "-C", str(local), "add", "a"])
-        utils_run_mod.run_checked(["git", "-C", str(local), "commit", "-m", "main commit"])
-        utils_run_mod.run_checked(["git", "-C", str(local), "branch", "-M", "main"])
-        utils_run_mod.run_checked(["git", "-C", str(local), "push", "origin", "main"])
-        utils_run_mod.run_checked(["git", "-C", str(local), "checkout", "-b", "feature"])
-        (local / "b").write_text("2\n")
-        utils_run_mod.run_checked(["git", "-C", str(local), "add", "b"])
-        utils_run_mod.run_checked(["git", "-C", str(local), "commit", "-m", "feature work"])
-        utils_run_mod.run_checked(["git", "-C", str(local), "push", "origin", "feature"])
-
-        # Pushed → reachable from refs/remotes/origin/feature → nothing absent.
-        assert git.commits_absent_from_all_remotes(str(local), "feature") == []
-
-    def test_local_only_commit_is_flagged(self, tmp_path: Path) -> None:
-        """A commit that was never pushed anywhere is reported (data loss risk)."""
-        bare = tmp_path / "remote.git"
-        utils_run_mod.run_checked(["git", "init", "--bare", str(bare)])
-        local = tmp_path / "local"
-        utils_run_mod.run_checked(["git", "clone", str(bare), str(local)])
-        for k, v in {"user.email": "t@x", "user.name": "t", "commit.gpgsign": "false"}.items():
-            utils_run_mod.run_checked(["git", "-C", str(local), "config", k, v])
-        (local / "a").write_text("1\n")
-        utils_run_mod.run_checked(["git", "-C", str(local), "add", "a"])
-        utils_run_mod.run_checked(["git", "-C", str(local), "commit", "-m", "main commit"])
-        utils_run_mod.run_checked(["git", "-C", str(local), "branch", "-M", "main"])
-        utils_run_mod.run_checked(["git", "-C", str(local), "push", "origin", "main"])
-        utils_run_mod.run_checked(["git", "-C", str(local), "checkout", "-b", "feature"])
-        (local / "b").write_text("2\n")
-        utils_run_mod.run_checked(["git", "-C", str(local), "add", "b"])
-        utils_run_mod.run_checked(["git", "-C", str(local), "commit", "-m", "never pushed"])
-
-        result = git.commits_absent_from_all_remotes(str(local), "feature")
-        assert len(result) == 1
-        assert "never pushed" in result[0]
-
-    def test_git_error_raises_instead_of_returning_empty(self, monkeypatch: pytest.MonkeyPatch) -> None:
-        """#706 fail-closed on a non-zero git exit.
-
-        A git error must NOT be read as "nothing unpushed" — it raises so the
-        data-loss guard refuses teardown.
-        """
-        monkeypatch.setattr(
-            utils_run_mod.subprocess,
-            "run",
-            lambda *_a, **_k: CompletedProcess([], 128, stdout="", stderr="fatal: bad revision"),
-        )
-        with pytest.raises(utils_run_mod.CommandFailedError):
-            git.commits_absent_from_all_remotes("/repo", "feature")
-
-    def test_missing_branch_raises_not_silently_empty(self, tmp_path: Path) -> None:
-        """An unknown branch makes ``git log`` exit 128 → must raise, not return []."""
-        bare = tmp_path / "remote.git"
-        utils_run_mod.run_checked(["git", "init", "--bare", str(bare)])
-        local = tmp_path / "local"
-        utils_run_mod.run_checked(["git", "clone", str(bare), str(local)])
-        for k, v in {"user.email": "t@x", "user.name": "t", "commit.gpgsign": "false"}.items():
-            utils_run_mod.run_checked(["git", "-C", str(local), "config", k, v])
-        (local / "a").write_text("1\n")
-        utils_run_mod.run_checked(["git", "-C", str(local), "add", "a"])
-        utils_run_mod.run_checked(["git", "-C", str(local), "commit", "-m", "main commit"])
-        utils_run_mod.run_checked(["git", "-C", str(local), "branch", "-M", "main"])
-        utils_run_mod.run_checked(["git", "-C", str(local), "push", "origin", "main"])
-
-        with pytest.raises(utils_run_mod.CommandFailedError):
-            git.commits_absent_from_all_remotes(str(local), "does-not-exist")
-
-
-def _init_dirty_repo(path: Path, *, marker: str = "tracked") -> None:
-    """A repo with one committed file edited dirty + one untracked file.
-
-    ``marker`` distinguishes two repos by filename so a hijacked ``git`` call
-    produces a visibly-wrong patch (the decoy's filenames, not the target's).
-    """
-    path.mkdir()
-    utils_run_mod.run_checked(["git", "init", "-q", "-b", "main", str(path)])
-    for k, v in {"user.email": "t@x", "user.name": "t", "commit.gpgsign": "false"}.items():
-        utils_run_mod.run_checked(["git", "-C", str(path), "config", k, v])
-    (path / f"{marker}.txt").write_text("base\n", encoding="utf-8")
-    utils_run_mod.run_checked(["git", "-C", str(path), "add", "-A"])
-    utils_run_mod.run_checked(["git", "-C", str(path), "commit", "-q", "-m", "initial"])
-    (path / f"{marker}.txt").write_text("base\nEDITED\n", encoding="utf-8")
-    (path / f"{marker}_new.txt").write_text("new content\n", encoding="utf-8")
-
-
-class TestFullWorktreeDiff:
-    """#835 — the captured patch must be plain-``git apply``-able everywhere."""
-
-    def test_patch_has_standard_prefixes_under_diff_noprefix_config(self, tmp_path: Path) -> None:
-        """Repo-local ``diff.noprefix=true`` must NOT strip ``a/``/``b/`` prefixes.
-
-        ``git diff`` honours the caller's git config; a user with
-        ``diff.noprefix=true`` (common; was set on the review machine) would
-        otherwise get a prefix-less patch that a plain ``git apply`` cannot
-        restore — total loss of the captured work, the exact #835 scenario.
-        The config is set REPO-LOCAL on purpose: the conftest HOME sandbox
-        masks real user git config, so a HOME-scoped setting would not exercise
-        this. RED on ``git diff HEAD --binary`` (no prefix flags); GREEN once
-        ``--src-prefix=a/ --dst-prefix=b/`` is forced.
-        """
-        repo = tmp_path / "repo"
-        _init_dirty_repo(repo)
-        utils_run_mod.run_checked(["git", "-C", str(repo), "config", "diff.noprefix", "true"])
-
-        patch_text = git.full_worktree_diff(str(repo))
-
-        assert "--- a/tracked.txt" in patch_text
-        assert "+++ b/tracked.txt" in patch_text
-
-        restore = tmp_path / "restore"
-        utils_run_mod.run_checked(["git", "init", "-q", "-b", "main", str(restore)])
-        (restore / "tracked.txt").write_text("base\n", encoding="utf-8")
-        env = {k: v for k, v in os.environ.items() if not k.startswith("GIT_")}
-        diff_file = tmp_path / "wt.diff"
-        diff_file.write_text(patch_text, encoding="utf-8")
-        # Plain ``git apply`` — the restore contract the docstring/BLUEPRINT promise.
-        utils_run_mod.run_checked(["git", "-C", str(restore), "apply", str(diff_file)], env=env)
-        assert (restore / "tracked.txt").read_text(encoding="utf-8") == "base\nEDITED\n"
-        assert (restore / "tracked_new.txt").read_text(encoding="utf-8") == "new content\n"
-
-    def test_ignores_poisoned_git_env_overrides(self, tmp_path: Path) -> None:
-        """A hijacking ``GIT_*`` env must not redirect the diff to another repo.
-
-        The inline pre-commit ``pytest`` hook runs under an outer ``git commit``
-        exporting ``GIT_DIR``/``GIT_WORK_TREE``. Inherited, they would point the
-        capture's ``git`` at the outer repo instead of the worktree it was
-        asked about. RED if the ``GIT_*`` strip is reverted.
-        """
-        target = tmp_path / "target"
-        _init_dirty_repo(target, marker="target")
-        decoy = tmp_path / "decoy"
-        _init_dirty_repo(decoy, marker="decoy")
-
-        with patch.dict(
-            os.environ,
-            {"GIT_DIR": str(decoy / ".git"), "GIT_WORK_TREE": str(decoy)},
-        ):
-            patch_text = git.full_worktree_diff(str(target))
-
-        # The patch must describe the TARGET's files. A leaked GIT_* would
-        # redirect the diff to the decoy repo — its filenames would appear and
-        # the target's would not.
-        assert "a/target.txt" in patch_text, patch_text
-        assert "target_new.txt" in patch_text, patch_text
-        assert "decoy" not in patch_text, patch_text
